builds to wasm with wasm-pack
