+++ helixflow-app/Cargo.toml
@@ -3,7 +3,17 @@ name = "helixflow-app"
 version = "0.1.0"
 edition = "2024"
 
+[lib]
+crate-type = ["cdylib", "rlib"]
+
 [dependencies]
 helixflow-surreal.workspace = true
 helixflow-slint.workspace = true
 slint.workspace = true
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+console_error_panic_hook.workspace = true
+console_log.workspace = true
+getrandom.workspace = true
+log.workspace = true
+wasm-bindgen.workspace = true
