+++ helixflow-app/index.html
@@ -1,0 +1,11 @@
+<html>
+    <body>
+        <!-- canvas required by the Slint runtime -->
+        <canvas id="canvas"></canvas>
+        <script type="module">
+            // import the generated file.
+            import init from "./pkg/helixflow_app.js";
+            init();
+        </script>
+    </body>
+</html>
